packages/core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diligence_core"
version = "0.1.0"
description = "Shared core utilities"
requires-python = ">=3.10"
dependencies=[
    "fastapi>=0.110,<1",
    "pydantic>=2,<3",
    "fastembed>=0.7.0,<0.7.5",
    "numpy>=2.4.1,<2.5",
    "pypdf>=6.7.0,<6.8",
    "qdrant-client>=1.8,<2",
    "google-genai>=1.60,<2",
    "langfuse>=3.14.0,<4.0.0",
    "supabase>=2.28.0,<2.29"
]

[project.optional-dependencies]
tokenizers=[
    "transformers>=5.0,<5.2",
    "sentencepiece>=0.2.0,<0.2.2"
]
test = [
  "pytest",
  "pytest-cov",
]


[tool.setuptools]
packages = ["diligence_core"]
